Improve and test various nullable ref annotations
